backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[binary]==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.8.5
web3==6.11.3
pytest==7.4.3
httpx==0.25.2 